Refactor headings into separate file
