+++ src/md/mod.rs
@@ -1,4 +1,5 @@
 mod code;
+mod headings;
 mod highlight;
 mod parse;
 mod title;
